Se agrega a app las variables de entorno y se crea la base de datos

+++ models/localidad.py
@@ -3,7 +3,7 @@ from db import db
 class LocalidadModel(db.Model):
     __tablename__ = "localidades"
 
-    id = db.Column(db.Integer(2), primary_key=True)
+    id = db.Column(db.Integer, primary_key=True)
     name = db.Column(db.String(80), unique=True, nullable=False)
     cp = db.Column(db.Integer, unique=False, nullable=False)
     id_muni = db.Column(db.Integer, db.ForeignKey("municipios.id"), unique=False, nullable=False)
